Handles `Timeouts`
In some cases we might instantiate an `EvenSource` object that never finds it's URL.  In these cases we respond back with a timeout.  The amount of time given to look for a URL is on `MockEventGlobals.setTimeout`, but should be instant.  Mostly because we're not traveling through a network.  These are all local objects.