Adds `MockEvent.evtSource` attribute
There is a one to one relationship between the `MockEvent` [aka handler] and the `EventSource` object.  The reason I added this attribute was for extra message passing.  In this specific case, I was trying to tell the end user via the EventSource.onerror method that the EventSource instance is closed and can't receive any more messages.  In the regular situation the end user would be blind to this fact.  Not completely convinced, this is up for debate.